SLM/requirements.txt
transformers==4.30.0
PyPDF2==3.0.1
nltk==3.8.1
pyyaml==6.0.1
torch==2.2.0
numpy==1.24.3
